Fixed type of data returned by the search #833

+++ web/App_Code/LcRest/UserSearchResult.cs
@@ -18,7 +18,7 @@ namespace LcRest
         public string businessName;
         public string jobTitleNameSingular;
         public string otherJobTitles;
-        public decimal distance;
+        public double distance;
 
         /// <summary>
         /// Same as in PublicUserProfile
